feat: adding more default configs
